SW-4761 Create NoOrgRouter and TerrawareRouter which each provide clean separation of their included views
