+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
     ports:
       - '5432:5432'
     volumes:
-      - '$HOME/docker/volumes/postgres:/var/lib/postgresql'
+      - '$HOME/docker/volumes/postgres/data:/var/lib/postgresql/data'
     environment:
       POSTGRES_DB: 'terraware'
       POSTGRES_PASSWORD: 'terraware'
